webservice2/requirements.txt
flask
cassandra-driver
redis
